Ping button added
