+++ src/pages/Home/PingTransaction/index.ts
@@ -1,0 +1,1 @@
+export * from './PingTransaction';
